+++ src/imio/helpers/security.py
@@ -65,3 +65,16 @@ def generate_password(length=10, digits=3, upper=2, lower=1, special=1, readable
     if not readable:
         password = sample(password, len(password))
     return "".join(password)
+
+
+def setup_logger(level=20):
+    """
+        When running "bin/instance run ...", the logger level is 30 (warn).
+        It is possible to set it to 20 (info) or 10 (debug).
+    """
+    log = logging.getLogger()
+    log.setLevel(level)
+    for handler in logging.root.handlers:
+        if handler.level == 30 and handler.formatter is not None:
+            handler.level = level
+            break